Extract While to separate struct
